Change Clean to go back to real original material
I found a way to adjust your addon so when the scene is cleaned they return to the materials they had before. That method is not elegant at all though, it relys soley on the material names and the materials that do still exist in the .blend to figure out which material to give which object when cleaning. This means it changes the names of materials, which probably breakes some stuff, though I have not discorved that myself when using it so far. The "XX_Original" materials are here not tagged as "Fake user", because they are actually not needed anymore since the really original Material is being tagged as fake user and the reaplied to the objects it has previously been applied to.
Also my formatter changed a bunch of files in meaningless ways.